Fhttp: add RegisterRemoteImage functionality.
This is a wrapper around manually having to download the image and
registering it. It registers an image based on it's URL, it will
download that image and add it to the PDF. It will however not add it
immediately to the page. This is done by calling `Image()` with the same
URL.