Add Barcode contribution package.
This package adds the ability to easily add a barcode to the PDF. This
barcode is generated with the github.com/boombuler/barcode package.

It adds wrappers for all the available barcode types through the
`Register()` and `Register<Type>()` methods. These register methods put
the barcode on the PDF, but not on the page. They will return a unique
key that should be used later on with the `Barcode()` method scale the
barcode and put it on the page.